% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot.R
\name{compare_plot}
\alias{compare_plot}
\title{Create an Interactive Scatter Plot for Comparing Variables}
\usage{
compare_plot(data, x, y, color, size = 5, alpha = 0.7, title = NULL)
}
\arguments{
\item{data}{A data frame containing the variables to plot}

\item{x}{Character string with the name of the variable to plot on the x-axis}

\item{y}{Character string with the name of the variable to plot on the y-axis}

\item{color}{Character string with the name of the variable to use for
coloring points}

\item{size}{Numeric value for the size of the points (default: 5)}

\item{alpha}{Numeric value for the transparency of the points (default: 0.7)}

\item{title}{Character string for the plot title. If NULL, a default title is
generated}
}
\value{
A plotly interactive scatter plot object
}
\description{
This function creates an interactive scatter plot using plotly to compare two
variables with points colored by a third variable.
}
\examples{
require(webshot)
compare_plot(movies, 
             "imdb_rating", "critics_score", "genre", 
              size = 7, alpha = 0.75, "Plot Title")

}
