% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var_name.R
\name{extract_var_name}
\alias{extract_var_name}
\title{Extract variable names safely}
\usage{
extract_var_name(var_expr)
}
\arguments{
\item{var_expr}{variable expression}
}
\value{
cleaned variable name
}
\description{
Convert any language/expression to character and extract just the variable name
}
