% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_name.R
\name{get_column_name}
\alias{get_column_name}
\title{Get column names from movies}
\usage{
get_column_name(var)
}
\arguments{
\item{var}{variable from UI}
}
\value{
character vector of column name from \code{movies} data
}
\description{
Get column names from movies
}
