% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny Application}
\usage{
launch_app(options = list(), run = "p", ...)
}
\arguments{
\item{options}{A list of options for the \code{shinyApp} function.}

\item{run}{A character string specifying the display type, default is \code{"p"}.}

\item{...}{Additional arguments passed to the UI function.}
}
\description{
This function initializes and launches the Shiny application with specified
options and settings.
}
