% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_vars.R
\name{mod_compare_vars_server}
\alias{mod_compare_vars_server}
\title{Server Logic for Variable Selection}
\usage{
mod_compare_vars_server(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A reactive list of selected variables and plot attributes.
}
\description{
This function handles the logic for variable selection, ensuring that
x and y variables are different.
}
\details{
When a user selects the same variable for both axes, the other
input is automatically updated to a different choice.
}
\section{Logging}{

Includes logging for variable selection changes and validation.
}

\seealso{
\code{\link[=mod_compare_vars_ui]{mod_compare_vars_ui()}}
}
