% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_waffle.R
\name{mod_counts_waffle_server}
\alias{mod_counts_waffle_server}
\title{Server Logic for Vertical Bar Plot}
\usage{
mod_counts_waffle_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of plot parameters.}
}
\value{
Renders a vertical bar plot based on selected parameters.
}
\description{
Server Logic for Vertical Bar Plot
}
\details{
This server function creates a waffle plot visualization using
the movie dataset. It filters data based on selected years and creates
a waffle chart representation of the selected categorical variable.
}
\section{Log Messages}{

Contains logging for errors and tracking critical rendering steps.
}

\seealso{
\code{\link[=mod_counts_waffle_ui]{mod_counts_waffle_ui()}}
}
