% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dist_box.R
\name{mod_dist_box_server}
\alias{mod_dist_box_server}
\title{Server Logic for Box Plot}
\usage{
mod_dist_box_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of plot parameters.}
}
\value{
Renders a box plot based on selected parameters.
}
\description{
Server Logic for Box Plot
}
\details{
This server function generates a box plot visualization using the
provided parameters in the reactive 'vals' list. It properly handles missing
values and applies aesthetic parameters like alpha and size.
}
\section{Associated UI Function}{

This server function should be used with the corresponding UI function
\code{mod_dist_box_ui()}.
}

