% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dist_tbl.R
\name{mod_dist_tbl_server}
\alias{mod_dist_tbl_server}
\title{Server Logic for Distribution Data Table}
\usage{
mod_dist_tbl_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of table parameters.}
}
\value{
Renders a data table based on selected parameters.
}
\description{
Server Logic for Distribution Data Table
}
\details{
This module creates a reactive data table that displays distribution
summaries based on the character and numeric variables selected from
the \code{vals()} reactive input.
}
\section{UI Function}{

Use with \code{mod_dist_tbl_ui()} to create the UI component for this module.
}

\seealso{
\code{\link[=mod_dist_tbl_ui]{mod_dist_tbl_ui()}}
}
