% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_server}
\alias{mod_report_server}
\title{Generate and download R Markdown report}
\usage{
mod_report_server(id, scatter_values, count_values, dist_values, award_values)
}
\arguments{
\item{id}{Module ID}

\item{scatter_values}{Scatter plot reactive values}

\item{count_values}{Count data reactive values}

\item{dist_values}{Distribution reactive values}

\item{award_values}{Awards reactive values}
}
\value{
None
}
\description{
Generate and download R Markdown report
}
