% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_ui}
\alias{mod_report_ui}
\title{Report download UI}
\usage{
mod_report_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for downloading report
}
\description{
Report download UI
}
