% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_server.R
\name{nav_server}
\alias{nav_server}
\title{Application Navigation Server Logic}
\usage{
nav_server(input, output)
}
\arguments{
\item{input, output}{Shiny server input and output objects.}
}
\value{
Server logic for the Shiny application with navigation.
}
\description{
Application Navigation Server Logic
}
\seealso{
\code{\link[=nav_ui]{nav_ui()}}
}
