% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_ui.R
\name{nav_ui}
\alias{nav_ui}
\title{User Interface for the Shiny Application}
\usage{
nav_ui()
}
\value{
A tagList containing the UI elements.
}
\description{
This function defines the user interface for the Shiny application,
integrating Bootstrap and a custom JavaScript file.
}
\seealso{
\code{\link[=nav_server]{nav_server()}}
}
