% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Create a Scatter Plot with Custom Axes and Color}
\usage{
scatter_plot(df, x_var, y_var, col_var)
}
\arguments{
\item{df}{A data frame or tibble containing the variables to be plotted.}

\item{x_var}{A string. The name of the variable to map to the x-axis.}

\item{y_var}{A string. The name of the variable to map to the y-axis.}

\item{col_var}{A string. The name of the variable to use for point color.}
}
\value{
A \code{ggplot} object representing the scatter plot.
}
\description{
Generates a \code{ggplot2} scatter plot from a data frame using variable names
supplied as strings for the x-axis, y-axis, and color aesthetic. This allows
for dynamic plotting in Shiny apps or other interactive environments.
}
\examples{
scatter_plot(mtcars, "mpg", "hp", "cyl")

}
